restructured directory, added and removed files
