+++ vite-project/src/App.tsx
@@ -1,10 +1,10 @@
 import './App.css';
-import './components/HomePage';
-import { HomePageRenderer } from './components/HomePage';
-import { TeamBuilder } from './components/TeamBuilder';
-import { TrainerPage } from './components/TrainerPage';
+import './pages/HomePage.tsx';
+import { HomePageRenderer } from './pages/HomePage.tsx';
+// import { TeamBuilder } from './components/TeamBuilder';
+import { TrainerPage } from './pages/TrainerPage.tsx';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import DisplayPokemon from './components/testComponents/DisplayPokemon.tsx';
+// import DisplayPokemon from './components/testComponents/DisplayPokemon.tsx';
 import PokemonTeamBuilder from './pages/PokemonTeamBuilder.tsx';
 // import PokemonTeamCreator from './components/pokemonComponents/PokemonTeamCreator.tsx';
 
@@ -15,9 +15,7 @@ function App() {
       <Routes>
         <Route path='/' element={<HomePageRenderer />}></Route>
         <Route path='/trainer' element={<TrainerPage />}></Route>
-        <Route path='/teambuilder' element={<TeamBuilder />}></Route>
-        <Route path='/displaypokemon' element={<DisplayPokemon />}></Route>
-        <Route path='/savepokemon' element={<PokemonTeamBuilder/>}></Route>
+        <Route path='/pokemonTeamBuilder' element={<PokemonTeamBuilder/>}></Route>
         {/* <Route path='/show' element = {<PokemonTeamCreator/>}></Route> */}
       </Routes>
     </Router>
