+++ vite-project/src/components/pokemonComponents/HeldItemList.tsx
