+++ vite-project/src/models/TrainerPageLogic.tsx
