+++ vite-project/src/pages/TrainerPage.css
